Leave Play for play.api.Play
Move play.api.Mode to its own file.  While its an important part of the
API, it shouldn't live above the actual Play object.
